project1/requirements.txt
flask==1.0.2
flask-login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
psycopg2==2.7.5
PyYAML==3.13